requirements.txt
django==4.2.13
Pillow==10.3.0
djangorestframework==3.15.2
drf-spectacular==0.27.2